+++ src/core/auth/options.ts
@@ -51,23 +51,6 @@ export type AuthTransportCallback =
   | ChannelAuthorizationCallback
   | UserAuthenticationCallback;
 
-export interface AuthOptionsT<AuthHandler> {
-  transport: 'ajax' | 'jsonp';
-  endpoint: string;
-  params?: any;
-  headers?: any;
-  paramsProvider?: () => any;
-  headersProvider?: () => any;
-  customHandler?: AuthHandler;
-}
-
-export declare type UserAuthenticationOptions = AuthOptionsT<
-  UserAuthenticationHandler
->;
-export declare type ChannelAuthorizationOptions = AuthOptionsT<
-  ChannelAuthorizationHandler
->;
-
 export interface InternalAuthOptions {
   transport: 'ajax' | 'jsonp';
   endpoint: string;
@@ -76,3 +59,18 @@ export interface InternalAuthOptions {
   paramsProvider?: () => any;
   headersProvider?: () => any;
 }
+
+export type CustomAuthOptions<AuthHandler> = {
+  customHandler: AuthHandler;
+};
+
+export type AuthOptionsT<AuthHandler> =
+  | InternalAuthOptions
+  | CustomAuthOptions<AuthHandler>;
+
+export declare type UserAuthenticationOptions = AuthOptionsT<
+  UserAuthenticationHandler
+>;
+export declare type ChannelAuthorizationOptions = AuthOptionsT<
+  ChannelAuthorizationHandler
+>;